+++ quadax/__init__.py
@@ -2,15 +2,13 @@
 
 from . import _version
 from .adaptive import adaptive_quadrature, quadcc, quadgk, quadts
+from .adjoint import AbstractAdjoint, DirectAdjoint, LeibnizAdjoint
 from .fixed_order import (
     AbstractQuadratureRule,
     ClenshawCurtisRule,
     GaussKronrodRule,
     NestedRule,
     TanhSinhRule,
-    fixed_quadcc,
-    fixed_quadgk,
-    fixed_quadts,
 )
 from .romberg import romberg, rombergts
 from .sampled import cumulative_simpson, cumulative_trapezoid, simpson, trapezoid
@@ -26,6 +24,9 @@ __all__ = [
     "GaussKronrodRule",
     "NestedRule",
     "TanhSinhRule",
+    "AbstractAdjoint",
+    "DirectAdjoint",
+    "LeibnizAdjoint",
     "romberg",
     "rombergts",
     "cumulative_simpson",